refactor(e2e): extract persistence/ package (checkpoint, result_writer, rehydrate) (#1940)
First slice of the e2e god-package decomposition. Moves checkpoint,
experiment_result_writer, and rehydrate modules into a new
scylla.persistence package. Leaves back-compat shims at the old paths
so existing external imports and tests keep working. Internal e2e/
modules now import from scylla.persistence.* directly, not through
the shim.

Test patches targeting module internals (e.g. patch("scylla.e2e.checkpoint.os.getpid"))
were updated to point at scylla.persistence.* — patching internals
across a re-export shim does not affect the source module.

Deferred slices tracked in #1940:
- orchestration/ (EvalOrchestrator, runner, tier_manager, scheduling)
- state_machines/ (experiment/tier/subtest/run FSMs — see #1942)
- judging/ (judge_runner, llm_judge)
- rerun/ (rerun, rerun_judges, regenerate)
- Cross-cutting types -> core/ (TokenStats — see #1937 Edge 3)
- import-linter CI gate (after all slices land)

Refs #1940
