+++ src/scylla/e2e/tier_state_machine.py
@@ -27,7 +27,7 @@ from typing import TYPE_CHECKING
 from scylla.e2e.models import TierState
 
 if TYPE_CHECKING:
-    from scylla.e2e.checkpoint import E2ECheckpoint
+    from scylla.persistence.checkpoint import E2ECheckpoint
 
 logger = logging.getLogger(__name__)
 
@@ -218,7 +218,7 @@ class TierStateMachine:
             ValueError: If no transition is defined for the current state
 
         """
-        from scylla.e2e.checkpoint import save_checkpoint
+        from scylla.persistence.checkpoint import save_checkpoint
 
         current = self.get_state(tier_id)
 
@@ -291,9 +291,9 @@ class TierStateMachine:
                     )
                     break
         except Exception as e:
-            from scylla.e2e.checkpoint import save_checkpoint
             from scylla.e2e.rate_limit import RateLimitError
             from scylla.e2e.shutdown import ShutdownInterruptedError
+            from scylla.persistence.checkpoint import save_checkpoint
 
             if isinstance(e, ShutdownInterruptedError):
                 # Ctrl+C interrupted this tier — leave it at CONFIG_LOADED (resumable)
