+++ src/scylla/persistence/__init__.py
@@ -1,0 +1,5 @@
+"""Persistence layer — checkpoint, result writing, rehydration.
+
+Extracted from scylla.e2e in #1940 as the first slice of the e2e
+god-package decomposition.
+"""
